fixes saves for nested objects + cleanup
The main thing here are the changes to .toJSON(), which no checks for the "actionName" on the options object.  This is cleaner than setting patch=true and relying on the attributes that get passed into .save().  It also allows for decoupling of what gets saved vs what gets sent to the server