[fix #35] Remove request promises from the store once they are resolve.
The point of storing request promises in the store is to deal with requests to an entity which already has an open request (e.g. we should just use the same request instead of creating a new one).  Keeping the requests in the store after the request has been resolved however, ends up a creating another caching layer.  (Which is further complicated by the fact that request promises store the raw json object and not the already hydrated bbSiren object)